[gui] split vm details file into multiple files
